refactor: move all tests to separate test folder for conveniency
