+++ src/tests/ErrorBoundary.test.tsx
