+++ src/tests/Search.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import Search from './Search';
+import Search from '../components/Search/Search';
 
 const mockSendSearchUp = vi.fn();
